--- client/src/components/List/List.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./List.css";
-
-export function List({ children }) {
-  return (
-    <div className="list-overflow-container">
-      <ul className="list-group">{children}</ul>
-    </div>
-  );
-}
