+++ client/src/components/List/index.js
@@ -1,2 +1,16 @@
-export * from "./List";
-export * from "./ListItem";
+import React from "react";
+import "./style.css";
+
+// This file exports both the List and ListItem components
+
+export function List({ children }) {
+  return (
+    <div className="list-overflow-container">
+      <ul className="list-group">{children}</ul>
+    </div>
+  );
+}
+
+export function ListItem({ children }) {
+  return <li className="list-group-item">{children}</li>;
+}
